justhink_server/requirements/prod.txt
Django==5.1.1
djangorestframework==3.15.2
python-dotenv==1.0.1
django_cors_headers==4.4.0
django_redis==5.4.0
daphne==4.1.2
channels==4.1.0
channels_redis==4.2.0
djangorestframework_simplejwt==5.3.1